Feat(Delete Task): Null checks, delete silently fails

+++ src/main/java/io/littlehorse/connector/service/KubernetesService.java
@@ -10,6 +10,7 @@ import io.littlehorse.connector.kubernetes.KubernetesUtils;
 
 import jakarta.enterprise.context.ApplicationScoped;
 
+import java.util.Objects;
 import java.util.Optional;
 
 @ApplicationScoped
@@ -31,12 +32,18 @@ public class KubernetesService {
      */
     public Optional<GenericKubernetesResource> get(
             final String apiVersion, final String kind, final String namespace, final String name) {
-        return Optional.ofNullable(Optional.ofNullable(namespace)
-                .map(nullableNamespace -> client.genericKubernetesResources(apiVersion, kind)
-                        .inNamespace(nullableNamespace)
-                        .withName(name))
-                .orElse(client.genericKubernetesResources(apiVersion, kind).withName(name))
-                .get());
+
+        final Resource<GenericKubernetesResource> resource = Objects.requireNonNull(
+                Optional.ofNullable(namespace)
+                        .map(nullableNamespace -> client.genericKubernetesResources(
+                                        apiVersion, kind)
+                                .inNamespace(nullableNamespace)
+                                .withName(name))
+                        .orElse(client.genericKubernetesResources(apiVersion, kind)
+                                .withName(name)),
+                "Provide a valid resource operation");
+
+        return Optional.ofNullable(resource.get());
     }
 
     /**
@@ -84,12 +91,15 @@ public class KubernetesService {
      * @param resource Any resource.
      * @return Result.
      */
-    public HasMetadata apply(final Resource<? extends HasMetadata> resource) {
+    private HasMetadata apply(final Resource<? extends HasMetadata> resource) {
+        final Resource<? extends HasMetadata> requiredResource =
+                Objects.requireNonNull(resource, "Provide a valid resource operation");
+
         try {
-            return resource.create();
+            return requiredResource.create();
         } catch (final KubernetesClientException e) {
             if (KubernetesUtils.isAlreadyExistsException(e)) {
-                return resource.update();
+                return requiredResource.update();
             }
             throw e;
         }
@@ -100,7 +110,7 @@ public class KubernetesService {
      *
      * @param resource Any resource.
      */
-    public void delete(final Resource<? extends HasMetadata> resource) {
-        resource.delete();
+    private void delete(final Resource<? extends HasMetadata> resource) {
+        if (resource != null) resource.delete();
     }
 }
